Consolidate and clean auth models
