+++ server/auth/github.js
@@ -1,7 +1,7 @@
 var passport = require('passport');
 var GitHubStrategy = require('passport-github2').Strategy;
 
-var User = require('../models/users-auth');
+var User = require('../models/Users');
 var config = require('./config');
 var init = require('./init');
